moving entirely to new log
