+++ lib/bootstrap.js
@@ -10,7 +10,7 @@ const path = require('path'),
   references = require('./services/references'),
   db = require('./services/db'),
   highland = require('highland');
-var log = require('./services/logger').setup({
+var log = require('./services/log').setup({
     file: __filename,
     action: 'bootstrap'
   }),
