+++ lib/control.js
@@ -1,7 +1,7 @@
 'use strict';
 
 var memoryLeakThreshold = 32768,
-  log = require('./services/logger').setup({
+  log = require('./services/log').setup({
     file: __filename
   });
 const _ = require('lodash'),
