+++ lib/services/db.js
@@ -1,25 +1,13 @@
-/**
- * Mostly just converts callbacks to Promises
- *
- * @module
- */
-
 'use strict';
-
-// for now, use memory.
 
 const _ = require('lodash'),
   jsonTransform = require('./../streams/json-transform'),
-  chalk = require('chalk'),
   Eventify = require('eventify'),
   validation = require('../validation'),
   promiseDefer = require('../utils/defer'),
   plugins = require('../plugins'),
   publishedVersionSuffix = '@published';
-let db = require('levelup')('whatever', { db: require('memdown') }),
-  log = require('./logger').setup({
-    file: __filename
-  });
+let db = require('levelup')('whatever', { db: require('memdown') });
 
 /**
  * Use ES6 promises
