+++ lib/services/notifications.js
@@ -4,7 +4,7 @@ const _ = require('lodash'),
   bluebird = require('bluebird'),
   contentType = 'Content-type',
   rest = require('../rest');
-var log = require('./logger').setup({
+var log = require('./log').setup({
   file: __filename
 });
