+++ lib/services/schedule.js
@@ -2,7 +2,7 @@
 
 var interval,
   intervalDelay = 50000 + Math.floor(Math.random() * 10000), // just under one minute
-  log = require('./logger').setup({
+  log = require('./log').setup({
     file: __filename
   });
 const _ = require('lodash'),
