+++ lib/services/upgrade.js
@@ -7,7 +7,7 @@ const _ = require('lodash'),
   files = require('../files'),
   path = require('path'),
   utils = require('../utils/components');
-var log = require('./logger').setup({
+var log = require('./log').setup({
   file: __filename,
   action: 'upgrade'
 });
